src/main/java/com/maikrantetasik/parkinglot/services/parsers/InputFileParser.java
package com.maikrantetasik.parkinglot.services.parsers;

import java.io.BufferedReader;
import java.io.FileReader;
import java.io.IOException;
import java.util.ArrayList;
import java.util.List;

public class InputFileParser implements FileParser {

    public String[] parse(String filePath) throws IOException {
        BufferedReader reader = new BufferedReader(new FileReader(filePath));
        List<String> lines = new ArrayList<>();

        String line;
        do {
            line = reader.readLine();
            if (line != null) {
                lines.add(line);
            }
        } while (line != null);
        reader.close();

        String[] result = new String[lines.size()];
        return lines.toArray(result);
    }
}
